Fixes a bug whereby akamai parameter configuration did not work as documented:
Changed __init__.py so that it would not error on a missing .akamai file prior to checking for its existence.

In addition, changed to not Raise an exception on a missing .akami file; instead, it prints a warning indicating that the necessary kwargs must be passed into PurgeRequest() init. This satisfies the documented ability to configure the purger via environment variables, a .akamai file, or via direct arguments to PurgeRequest()